Updated packages. Fixed watcher. Fixed destination folder. Bundled example gulpfile.
